fix bug in scale pyramid
Since we are hacking the scale pyramid into the LocalVolume that is unaware of any scaling, neuroglancer
will only attempt to fetch isotropic downsampled keys. If your downsampling is even slightly off the desired keys won't be available.
Fortunately the LocalVolume does do some in-memory downsampling, so we can simply pass on the closest scale level and let the LocalVolume handle the rest